backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ML and Embedding Libraries
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0

# Vector Database
qdrant-client==1.7.0

# S3 Storage
boto3==1.34.10

# Clustering and Numerical
scipy==1.11.4
numpy==1.26.2

# File Handling
python-multipart==0.0.6
aiofiles==23.2.1
